# openapi.yaml
openapi: 3.0.3
info:
  title: Global Real Estate Dashboard API
  version: 1.0.0
  description: REST API for managing global real estate listings, viewings, documents, and market data.
servers:
  - url: /api/v1
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterInput" }
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data: { $ref: "#/components/schemas/AuthPayload" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data: { $ref: "#/components/schemas/AuthPayload" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }

    put:
      tags: [Auth]
      summary: Update current user profile
      operationId: updateMe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ApiResponse:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        data: {}
        meta: { type: object }
        message: { type: string }
        error: { type: string }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error: { type: string }
...
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        created_at: { type: string, format: date-time }
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
    AuthPayload:
      type: object
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }